# docs/openapi/openapi.yaml
...
openapi: 3.0.4
...
components:
  schemas:
...
    SendFriendInvitationDTO:
      required:
      - receiverEmail
      type: object
      properties:
        receiverEmail:
          pattern: \S
          type: string
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
...
  /api/friends/invitations:
    post:
      operationId: FriendInvitationResource_sendInvitation
      tags:
      - "User: Friends"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendFriendInvitationDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
        "400":
          description: Bad Request
      summary: Send Invitation
      security:
      - jwt:
        - USER
        - ADMIN
  /api/friends/invitations/received:
    get:
      operationId: FriendInvitationResource_getReceivedInvitations
      tags:
      - "User: Friends"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Get Received Invitations
      security:
      - jwt:
        - USER
        - ADMIN
  /api/friends/invitations/sent:
    get:
      operationId: FriendInvitationResource_getSentInvitations
      tags:
      - "User: Friends"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Get Sent Invitations
      security:
      - jwt:
        - USER
        - ADMIN
  /api/friends/invitations/{invitationId}/accept:
    put:
      operationId: FriendInvitationResource_acceptInvitation
      tags:
      - "User: Friends"
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Accept Invitation
      security:
      - jwt:
        - USER
        - ADMIN
  /api/friends/invitations/{invitationId}/cancel:
    put:
      operationId: FriendInvitationResource_cancelInvitation
      tags:
      - "User: Friends"
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Cancel Invitation
      security:
      - jwt:
        - USER
        - ADMIN
  /api/friends/invitations/{invitationId}/reject:
    put:
      operationId: FriendInvitationResource_rejectInvitation
      tags:
      - "User: Friends"
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Reject Invitation
      security:
      - jwt:
        - USER
        - ADMIN
...
info:
  title: GeoPulse API
  version: 1.34.1
  description: GeoPulse backend REST API documentation.
  contact:
    name: tess1o
    email: kerriden1@gmail.com
servers:
- url: http://localhost:8080